gillajabi/requirements.txt
absl-py==0.13.0
alabaster==0.7.12
altair==4.1.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
appnope==0.1.2
argcomplete==1.12.3
argon2-cffi==20.1.0
astor==0.8.1
attrs==21.2.0
audioread==2.1.9
autoflake==1.4
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
bandit==1.7.0
base58==2.1.0
beautifulsoup4==4.6.0
black==20.8b1
bleach==4.0.0
blinker==1.4
Bottleneck==1.3.2
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
ci-sdr==0.0.0
click==7.1.2
colorama==0.4.4
ConfigArgParse==1.5.2
configparser==5.0.2
coverage==5.5
ctc-segmentation==1.6.7
cycler==0.10.0
Cython==0.29.24
darglint==1.8.0
dataclasses==0.6
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
Distance==0.1.3
distlib==0.3.2
doc8==0.8.1
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.17.1
editdistance==0.5.2
einops==0.3.0
emoji==1.4.2
entrypoints==0.3
eradicate==1.0
espnet==0.10.1
espnet-tts-frontend==0.0.3
fairseq==0.10.2
fastwer==0.1.3
ffmpeg==1.4
filelock==3.0.12
flake8==3.9.2
flake8-2020==1.6.0
flake8-bandit==2.1.2
flake8-broken-line==0.2.1
flake8-bugbear==19.8.0
flake8-builtins==1.5.3
flake8-commas==2.0.0
flake8-comprehensions==3.6.1
flake8-debugger==3.2.1
flake8-docstrings==1.6.0
flake8-eradicate==0.3.0
flake8-fixme==1.1.1
flake8-isort==3.0.1
flake8-logging-format==0.6.0
flake8-mutable==1.2.0
flake8-polyfill==1.0.2
flake8-quotes==2.1.2
flake8-rst-docstrings==0.0.12
flake8-string-format==0.2.3
flake8-variables-names==0.0.3
future==0.18.2
g2p-en==2.1.0
g2pK==0.9.4
gitdb==4.0.7
GitPython==3.1.18
google-api-core==1.31.2
google-auth==1.35.0
google-auth-oauthlib==0.4.5
google-cloud-core==2.0.0
google-cloud-speech==1.3.2
google-cloud-storage==1.42.0
google-crc32c==1.1.2
google-resumable-media==2.0.0
googleapis-common-protos==1.53.0
grpcio==1.39.0
h5py==3.4.0
huggingface-hub==0.0.12
humanfriendly==9.2
hydra-core==1.1.1
idna==3.2
imagesize==1.2.0
importlib-metadata==4.6.4
importlib-resources==5.2.2
inflect==5.3.0
iniconfig==1.1.1
ipykernel==6.2.0
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==4.3.21
jaconv==0.3
jamo==0.4.1
jamotools==0.1.10
jedi==0.18.0
Jinja2==3.0.1
jiwer==2.2.0
joblib==1.0.1
JPype1==1.3.0
Js2Py==0.71
jsonschema==3.2.0
jupyter-client==7.0.1
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kaldiio==2.17.2
kenlm @ https://github.com/kpu/kenlm/archive/master.zip
kiwisolver==1.3.1
konlpy @ git+https://github.com/e9t/konlpy.git@21bc951bb54e4e875edc57b25fa52f70ac2a293e
korean-romanizer==0.22
kss==2.5.1
librosa==0.8.1
llvmlite==0.37.0
lxml==4.6.3
marisa-trie==0.7.7
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.1.0
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
nara-wpe==0.0.7
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.2
notebook==6.4.3
numba==0.54.0
numpy==1.20.3
oauthlib==3.1.1
omegaconf==2.1.1
packaging==21.0
pandas==1.3.2
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.9.0
pathtools==0.1.2
pbr==5.6.0
pep8-naming==0.9.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pipreqs==0.4.10
pipwin==0.5.1
platformdirs==2.2.0
pluggy==0.13.1
pooch==1.5.0
pororo==0.4.2
portalocker==2.3.1
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pybind11==2.7.0
pycodestyle==2.7.0
pycparser==2.20
pydeck==0.6.2
pydocstyle==5.1.1
pydub==0.25.1
pyflakes==2.3.1
Pygments==2.10.0
pyjsparser==2.7.1
pyparsing==2.4.7
pypinyin==0.42.0
PyPrind==2.11.3
pyrsistent==0.18.0
pySmartDL==1.3.4
PySocks==1.7.1
pytest==6.2.4
pytest-cov==2.12.1
python-dateutil==2.8.2
python-dev-tools==2020.9.10
python-Levenshtein==0.12.2
python-mecab-ko==1.0.11
pytorch-wpe==0.0.1
pytz==2021.1
pyupgrade==2.24.0
pyworld==0.3.0
PyYAML==5.4.1
pyzmq==22.2.1
regex==2021.8.21
requests==2.26.0
requests-oauthlib==1.3.0
resampy==0.2.2
restructuredtext-lint==1.3.2
rsa==4.7.2
sacrebleu==2.0.0
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.1
Send2Trash==1.8.0
sentence-transformers==2.0.0
sentencepiece==0.1.86
sentry-sdk==1.3.1
shortuuid==1.0.1
six==1.16.0
smmap==4.0.0
snowballstemmer==2.1.0
sounddevice==0.4.2
SoundFile==0.10.3.post1
soupsieve==2.2.1
Sphinx==2.4.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stevedore==3.4.0
streamlit==0.87.0
subprocess32==3.5.4
tabulate==0.8.9
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
terminado==0.11.1
testfixtures==6.18.1
testpath==0.5.0
threadpoolctl==2.2.0
tokenize-rt==4.1.0
tokenizers==0.10.3
toml==0.10.2
toolz==0.11.1
torch==1.6.0
torch-complex==0.2.1
torchvision==0.7.0
tornado==6.1
tox==3.24.3
tox-travis==0.12
tqdm==4.62.2
traitlets==5.0.5
transformers==4.9.2
tweepy==3.10.0
typed-ast==1.4.3
typeguard==2.12.1
typing-extensions==3.10.0.0
tzlocal==3.0
Unidecode==1.2.0
urllib3==1.26.6
validators==0.18.2
virtualenv==20.7.2
wandb==0.12.0
wavio==0.0.4
wcwidth==0.2.5
webencodings==0.5.1
wemake-python-styleguide==0.14.1
Werkzeug==2.0.1
wget==3.2
Whoosh==2.7.4
widgetsnbextension==3.5.1
word2word==1.0.0
yarg==0.1.9
zipp==3.5.0